backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
distlib==0.3.2
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
filelock==3.0.12
idna==2.10
Pillow==8.2.0
pycparser==2.20
pytz==2021.1
pyjwt-2.1.0
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.7
